Remove Provenance.cpp/h.
