+++ bsan-pass/CMakeLists.txt
@@ -12,13 +12,11 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${LLVM_CXXFLAGS}")
 set(BSAN_SOURCES
   BorrowSanitizer.cpp
   BorrowSanitizerPass.cpp
-  Provenance.cpp
   Retag.cpp
 )
 
 set(BSAN_HEADERS
   BorrowSanitizerPass.h
-  Provenance.h
   Retag.h
 )
